prepare to implement other strategies
While working on this approach I had the following insight. Basically we are working with trees of cells where each child of a node contains all not determined siblings in a certain block.

When the effective candidates of the root, i.e. the difference of the candidates of the root with the union of the candidates of the effective candidates of all the children, is a singleton we have a suggestion

A strategy amounts to iterating over a stream of candidates trees, and seeing if the root has an effective candidate set of size 1

Next commits will implements the ideas here